UPSTREAM: <carry>: v1beta3 scc volumes support
